renamed card component
